requirements.txt
gradio>=4.36.1
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.3.0
rembg>=2.0.57
imageio-ffmpeg>=0.5.1
